refactor(utils): remove unused files and code, consolidate DATA_FETCH_HOOKS, and rename import-utils to imports
