+++ codemods/v4/scripts/template-compilation-changes.ts
@@ -1,7 +1,7 @@
 import type { SgRoot, Edit } from "codemod:ast-grep";
 import type TS from "codemod:ast-grep/langs/typescript";
 import { hasContent } from "../utils/index.ts";
-import { ensureImport } from "../utils/import-utils.ts";
+import { ensureImport } from "../utils/imports.ts";
 
 async function transform(root: SgRoot<TS>): Promise<string | null> {
   const rootNode = root.root();
