+++ codemods/v4/tests/import-utils/case-6-type-to-runtime-test.ts
@@ -1,6 +1,6 @@
 import type { SgRoot } from "codemod:ast-grep";
 import type TSX from "codemod:ast-grep/langs/tsx";
-import { ensureImport } from "../../utils/import-utils.ts";
+import { ensureImport } from "../../utils/imports.ts";
 
 /**
  * CASE 6: REPLACE - Type import exists, replace with runtime import
