+++ codemods/v4/utils/ast-utils.ts
@@ -1,13 +1,23 @@
-import type { SgRoot, SgNode, Edit } from "codemod:ast-grep";
+import type { SgRoot, SgNode, Edit, TypesMap } from "codemod:ast-grep";
 
 /**
  * Core AST utilities for codemods
  */
 
 /**
+ * Common Nuxt data fetching hooks
+ */
+export const DATA_FETCH_HOOKS = [
+  "useAsyncData",
+  "useFetch",
+  "useLazyAsyncData",
+  "useLazyFetch",
+] as const;
+
+/**
  * Quick check if file contains specific content before processing
  */
-export function hasContent<T extends Record<string, any>>(
+export function hasContent<T extends TypesMap>(
   root: SgRoot<T>,
   searchText: string
 ): boolean {
@@ -17,7 +27,7 @@ export function hasContent<T extends Record<string, any>>(
 /**
  * Check if file contains any of the specified content
  */
-export function hasAnyContent<T extends Record<string, any>>(
+export function hasAnyContent<T extends TypesMap>(
   root: SgRoot<T>,
   searchTexts: readonly string[]
 ): boolean {
@@ -28,7 +38,7 @@ export function hasAnyContent<T extends Record<string, any>>(
 /**
  * Apply edits and return result, or null if no changes
  */
-export function applyEdits<T extends Record<string, any>>(
+export function applyEdits<T extends TypesMap>(
   rootNode: SgNode<T>,
   edits: Edit[]
 ): string | null {
@@ -39,36 +49,9 @@ export function applyEdits<T extends Record<string, any>>(
 }
 
 /**
- * Find function calls with multiple quote styles
- */
-export function findFunctionCalls<T extends Record<string, any>>(
-  rootNode: SgNode<T>,
-  functionName: string,
-  ...args: string[]
-): SgNode<T>[] {
-  const results: SgNode<T>[] = [];
-  const argPattern = args.length > 0 ? args.join(", ") : "$$$ARGS";
-
-  // Try both single and double quotes for string literals
-  const patterns = [
-    `${functionName}(${argPattern})`,
-    `await ${functionName}(${argPattern})`,
-  ];
-
-  for (const pattern of patterns) {
-    const calls = rootNode.findAll({
-      rule: { pattern },
-    });
-    results.push(...calls);
-  }
-
-  return results;
-}
-
-/**
  * Find function calls with specific first argument (handles quote variations)
  */
-export function findFunctionCallsWithFirstArg<T extends Record<string, any>>(
+export function findFunctionCallsWithFirstArg<T extends TypesMap>(
   rootNode: SgNode<T>,
   functionName: string,
   firstArg: string
@@ -77,8 +60,8 @@ export function findFunctionCallsWithFirstArg<T extends Record<string, any>>(
 
   // Handle both quote styles
   const patterns = [
-    `${functionName}('${firstArg}', $$$REST)`,
-    `${functionName}("${firstArg}", $$$REST)`,
+    `${functionName}('${firstArg}', $CALLBACK)`,
+    `${functionName}("${firstArg}", $CALLBACK)`,
   ];
 
   for (const pattern of patterns) {
@@ -94,7 +77,7 @@ export function findFunctionCallsWithFirstArg<T extends Record<string, any>>(
 /**
  * Replace text in node using regex - returns edit or null
  */
-export function replaceInNode<T extends Record<string, any>>(
+export function replaceInNode<T extends TypesMap>(
   node: SgNode<T>,
   searchRegex: RegExp,
   replacement: string
@@ -105,23 +88,4 @@ export function replaceInNode<T extends Record<string, any>>(
     return node.replace(newText);
   }
   return null;
-}
-
-/**
- * Find nodes matching multiple patterns
- */
-export function findWithPatterns<T extends Record<string, any>>(
-  rootNode: SgNode<T>,
-  patterns: string[]
-): SgNode<T>[] {
-  const results: SgNode<T>[] = [];
-
-  for (const pattern of patterns) {
-    const matches = rootNode.findAll({
-      rule: { pattern },
-    });
-    results.push(...matches);
-  }
-
-  return results;
 }
